Add recruitment to prefab creator

+++ TruckLib/ScsMap/Service.cs
@@ -20,7 +20,7 @@ namespace TruckLib.ScsMap
         public ServiceType ServiceType
         {
             get => (ServiceType)Flags.GetByte(0);
-            set => Flags.SetByte(0, value);
+            set => Flags.SetByte(0, (byte)value);
         }
 
         public static Service Add(IItemContainer map, Prefab parent, Vector3 position)
